requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pandas==2.1.3
numpy==1.25.2
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.17.0
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.0
pydantic-settings==2.1.0
streamlit==1.28.1
jinja2==3.1.2
python-multipart==0.0.6
statsmodels==0.14.0
scipy==1.11.3
requests==2.31.0
aiohttp==3.9.1
astropy==5.3.4
beautifulsoup4==4.12.2
